HANA_HOMEPLANNER_CRAWLER/producer/requirements.txt
apscheduler==3.10.4
kafka-python==2.0.2
redis==5.0.1
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0